% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_img.R
\name{df_to_img}
\alias{df_to_img}
\title{Convert data frame of pixel values to image}
\usage{
df_to_img(df)
}
\arguments{
\item{df}{A data frame with columns:
\itemize{
\item x,y: pixel coordinates (required)
\item Either:
\itemize{
\item R,G,B: RGB values for color images (0-1)
\item I: intensity value for grayscale images (0-1)
}
}}
}
\value{
A magick image object
}
\description{
Convert data frame of pixel values to image
}
