% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_blank.R
\name{img_blank}
\alias{img_blank}
\title{Create a blank square image}
\usage{
img_blank(width = 200, height = NULL, color = "black")
}
\arguments{
\item{color}{Character. Color name or hex code}

\item{size}{Integer. Width and height of the square image}
}
\value{
A magick-image object
}
\description{
Create a blank square image
}
