% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_color_distribution.R
\name{img_color_distribution}
\alias{img_color_distribution}
\title{Get color distribution from an image}
\usage{
img_color_distribution(img, max_colors = 20, method = c("quantize", "kmeans"))
}
\arguments{
\item{img}{A magick-image object}

\item{max_colors}{Integer. Maximum number of colors to include in the distribution (default: 20)}

\item{method}{Character. Method to use for color reduction if needed: "quantize" (default) or "kmeans"}
}
\value{
A tibble with columns:
\itemize{
\item color: Hex color code
\item count: Number of pixels with this color
\item distribution: Proportion of pixels with this color
}
}
\description{
Get color distribution from an image
}
