% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_has_more_than_n_colors.R
\name{img_has_more_than_n_colors}
\alias{img_has_more_than_n_colors}
\title{Quickly check if an image has more than n colors}
\usage{
img_has_more_than_n_colors(img, n = 20)
}
\arguments{
\item{img}{A magick-image object}

\item{n}{Number of colors threshold (default: 20)}
}
\value{
Logical. TRUE if image has more than n colors, FALSE otherwise
}
\description{
Quickly check if an image has more than n colors
}
