% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_mask.R
\name{img_mask}
\alias{img_mask}
\title{Create an image mask}
\usage{
img_mask(
  size,
  shape = c("circle", "rounded", "frame", "diamond", "oval"),
  radius = NULL
)
}
\arguments{
\item{size}{Integer. Size of the square mask}

\item{shape}{Character. Shape of the mask ("circle", "rounded", "frame", "diamond", "oval")}

\item{radius}{Numeric. Shape modifier (0-1):
\itemize{
\item circle: ignored
\item rounded: corner radius as proportion
\item frame: frame thickness as proportion
\item diamond: diamond size as proportion (1=smallest)
\item oval: height/width ratio (1 = circle)
}}
}
\value{
A magick-image object
}
\description{
Create an image mask
}
