% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_quantize.R
\name{img_quantize}
\alias{img_quantize}
\title{Quantize an image to reduce the number of colors}
\usage{
img_quantize(
  img,
  colors = 16,
  colorspace = "sRGB",
  dither = FALSE,
  treedepth = 0
)
}
\arguments{
\item{img}{A magick-image object}

\item{colors}{Integer. Number of colors to quantize the image to (default: 16)}

\item{colorspace}{Character. Colorspace to use for quantization:
"sRGB", "gray", "HCL", "HSB", "HSL", "Lab", "RGB" (default: "sRGB")}

\item{dither}{Logical. Whether to apply dithering (default: FALSE)}

\item{treedepth}{Integer. Maximum depth of the quantization tree (default: 0 for auto)}
}
\value{
A magick-image object with reduced colors
}
\description{
Quantize an image to reduce the number of colors
}
