% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_rotate.R
\name{img_rotate}
\alias{img_rotate}
\title{Rotate an image}
\usage{
img_rotate(img, degrees = 90)
}
\arguments{
\item{img}{A magick-image object}

\item{degrees}{Numeric. Rotation angle in degrees (clockwise)}

\item{background}{Character. Color to use for background when rotating creates empty space. Default "transparent"}

\item{expand}{Logical. If TRUE, expands the output image to hold the entire rotated image. Default TRUE}
}
\value{
A magick-image object
}
\description{
Rotate an image
}
