% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_sample.R
\name{img_sample}
\alias{img_sample}
\title{Create sample images for testing and demonstration}
\usage{
img_sample(
  type = c("windows", "checkerboard", "circles", "gradient", "stripes", "color_bars",
    "grayscale_steps", "bw_pattern"),
  width = 200,
  height = NULL,
  bg_color = "white",
  colors = NULL,
  n_shapes = 4
)
}
\arguments{
\item{type}{Character. Type of sample image to create: "windows", "checkerboard", "circles",
"gradient", "stripes", "color_bars", "grayscale_steps", "bw_pattern"}

\item{width}{Integer. Width of the image in pixels (default: 200)}

\item{height}{Integer. Height of the image in pixels (default: same as width)}

\item{bg_color}{Character. Background color (default: "white")}

\item{colors}{Character vector. Colors to use for the sample (will be recycled if needed)}

\item{n_shapes}{Integer. Number of shapes to draw for applicable sample types (default: 4)}
}
\value{
A magick-image object
}
\description{
Create sample images for testing and demonstration
}
