% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_to_df.R
\name{img_to_df}
\alias{img_to_df}
\title{Convert image to data frame of pixel values}
\usage{
img_to_df(img)
}
\arguments{
\item{img}{A magick-image object}
}
\value{
A data frame with columns:
\itemize{
\item x,y: pixel coordinates
\item R,G,B: RGB values for color images (0-1)
\item I: intensity value for grayscale images (0-1)
}
}
\description{
Convert image to data frame of pixel values
}
