% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imghelper-package.R
\docType{package}
\name{imghelper-package}
\alias{imghelper}
\alias{imghelper-package}
\title{imghelper: Image manipulation helpers}
\description{
Image manipulation helpers
}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

Other contributors:
\itemize{
  \item Datasketch [copyright holder, funder]
}

}
\keyword{internal}
